修改成Unity 设计，Player待检查，接下来处理TrackView，先生成轨道

+++ LuneRun_Unity/Assets/Scripts/com/playchilla/runner/player/IPlayerListener.cs
@@ -2,6 +2,6 @@ namespace com.playchilla.runner.player
 {
     public interface IPlayerListener
     {
-        void onLand(double impact);
+        void onLand(float impact);
     }
 }
